requirements.txt
astroid==1.6.4
autopep8==1.3.5
blessings==1.6.1
bpython==0.17.1
certifi==2018.4.16
chardet==3.0.4
curtsies==0.3.0
docopt==0.6.2
flake8==3.5.0
greenlet==0.4.13
grpcio==1.12.0
grpcio-tools==1.12.0
h2==3.0.1
hpack==3.0.0
hyperframe==5.1.0
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mypy==0.600
pep8==1.7.1
pipreqs==0.4.9
pluggy==0.6.0
protobuf==3.5.2.post1
py==1.5.3
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.9.1
requests==2.18.4
rope==0.10.7
six==1.11.0
tox==3.0.0
typed-ast==1.1.0
typing==3.6.4
urllib3==1.22
virtualenv==16.0.0
wcwidth==0.1.7
wrapt==1.10.11
yapf==0.22.0
yarg==0.1.9
